Move module setup to conftest.py
